requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
-e git://github.com/jsocol/django-adminplus@5520e8417f195352d80db86c9020d58e8407d585#egg=django_adminplus
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.25.2
